refactor and edits after review

+++ test/chain/test_chain_import_export.py
@@ -9,7 +9,7 @@ from core.composer.chain import Chain
 from core.composer.node import PrimaryNode, SecondaryNode
 from core.models.data import InputData
 from utilities.synthetic.chain_template_new import ChainTemplate, extract_subtree_root
-from test.chain.test_atomized_chain import create_chain_with_several_chain_models_nested
+from test.chain.test_atomized_model import create_chain_with_several_nested_atomized_model
 
 
 @pytest.fixture(scope="session", autouse=True)
@@ -19,7 +19,6 @@ def creation_model_files_before_after_tests(request):
 
     create_json_models_files()
     request.addfinalizer(delete_json_models_files)
-    request.addfinalizer(delete_folder_with_nested_model_json)
 
 
 def create_json_models_files():
@@ -36,35 +35,40 @@ def create_json_models_files():
     chain_empty.save_chain("data/test_empty_chain_convert_to_json.json")
 
 
-def delete_folder_with_nested_model_json():
+def delete_json_models_files():
     """
-    Delete folders.
+    Deletes files and folders created during testing.
     """
-    folders_name = ["data/chain_model_1",
-                    "data/check_load_model_chain_1"]
+
+    folders_name = ["data/atomized_model_1"]
 
     for folder_path in folders_name:
         dir_path = os.path.abspath(folder_path)
-        shutil.rmtree(dir_path)
-
-
-def delete_json_models_files():
-    """
-    Delete JSON's files.
-    """
-    if os.path.isfile("data/test_fitted_chain_convert_to_json.json"):
-        with open("data/test_fitted_chain_convert_to_json.json", 'r') as json_file:
-            chain_fitted_object = json.load(json_file)
-
-        delete_fitted_models(chain_fitted_object)
+        if os.path.exists(dir_path):
+            shutil.rmtree(dir_path)
 
     files_name = ["data/test_fitted_chain_convert_to_json.json",
                   "data/test_empty_chain_convert_to_json.json",
-                  "data/test_chain_convert_to_json.json"]
+                  "data/test_chain_convert_to_json.json",
+                  "data/1.json",
+                  "data/2.json",
+                  "data/3.json",
+                  "data/4.json",
+                  "data/5.json",
+                  "data/6.json",
+                  "data/7.json",
+                  "data/8.json",
+                  "data/atomized_model_1.json"]
 
     for file in files_name:
-        if os.path.isfile(file):
-            os.remove(file)
+        full_path = os.path.abspath(file)
+
+        if os.path.isfile(full_path):
+            with open(full_path, 'r') as json_file:
+                chain_fitted_object = json.load(json_file)
+
+            delete_fitted_models(chain_fitted_object)
+            os.remove(full_path)
 
 
 def delete_fitted_models(chain):
@@ -73,11 +77,13 @@ def delete_fitted_models(chain):
 
     :param chain: chain which model's need to delete
     """
-    root_node = chain['nodes'][0]
-    if 'trained_model_path' in root_node:
-        model_path = root_node['trained_model_path']
-        dir_path = os.path.dirname(os.path.abspath(model_path))
-        shutil.rmtree(dir_path)
+    if chain['nodes']:
+        root_node = chain['nodes'][0]
+        if 'trained_model_path' in root_node:
+            model_path = root_node['trained_model_path']
+            if model_path is not None and os.path.exists(model_path):
+                dir_path = os.path.dirname(os.path.abspath(model_path))
+                shutil.rmtree(dir_path)
 
 
 def create_chain() -> Chain:
@@ -125,6 +131,26 @@ def create_fitted_chain() -> Chain:
     return chain
 
 
+def create_four_depth_chain():
+    knn_node = PrimaryNode('knn')
+    lda_node = PrimaryNode('lda')
+    xgb_node = PrimaryNode('xgboost')
+    logit_node = PrimaryNode('logit')
+
+    logit_node_second = SecondaryNode('logit', nodes_from=[knn_node, lda_node])
+    xgb_node_second = SecondaryNode('xgboost', nodes_from=[logit_node])
+
+    qda_node_third = SecondaryNode('qda', nodes_from=[xgb_node_second])
+    knn_node_third = SecondaryNode('knn', nodes_from=[logit_node_second, xgb_node])
+
+    knn_root = SecondaryNode('knn', nodes_from=[qda_node_third, knn_node_third])
+
+    chain = Chain()
+    chain.add_node(knn_root)
+
+    return chain
+
+
 def test_export_chain_to_json_correctly():
     chain = create_chain()
     json_actual = chain.save_chain("data/1.json")
@@ -132,7 +158,6 @@ def test_export_chain_to_json_correctly():
     with open("data/test_chain_convert_to_json.json", 'r') as json_file:
         json_expected = json.load(json_file)
 
-    os.remove("data/1.json")
     assert json_actual == json.dumps(json_expected)
 
 
@@ -150,13 +175,11 @@ def test_chain_template_to_json_correctly():
 def test_import_json_to_chain_correctly():
     chain = Chain()
     chain.load_chain("data/test_chain_convert_to_json.json")
-    json_actual = chain.save_chain("data/1.json")
+    json_actual = chain.save_chain("data/2.json")
 
     chain_expected = create_chain()
-    json_expected = chain_expected.save_chain("data/2.json")
+    json_expected = chain_expected.save_chain("data/3.json")
 
-    os.remove("data/1.json")
-    os.remove("data/2.json")
     assert json.dumps(json_actual) == json.dumps(json_expected)
 
 
@@ -176,12 +199,11 @@ def test_import_json_template_to_chain_correctly():
 def test_import_json_to_fitted_chain_correctly():
     chain = Chain()
     chain.load_chain("data/test_fitted_chain_convert_to_json.json")
-    json_actual = chain.save_chain("data/1.json")
+    json_actual = chain.save_chain("data/4.json")
 
     with open("data/test_fitted_chain_convert_to_json.json", 'r') as json_file:
         json_expected = json.load(json_file)
 
-    os.remove("data/1.json")
     assert json_actual == json.dumps(json_expected)
 
 
@@ -218,34 +240,15 @@ def test_export_import_for_one_chain_object_correctly():
     and the last command will rewrite the chain object correctly.
     """
     chain_fitted = create_fitted_chain()
-    json_first = chain_fitted.save_chain("data/2.json")
+    json_actual = chain_fitted.save_chain("data/5.json")
 
     chain_fitted_after = create_chain()
-    chain_fitted_after.save_chain("data/1.json")
-    chain_fitted_after.load_chain("data/2.json")
+    chain_fitted_after.save_chain("data/6.json")
+    chain_fitted_after.load_chain("data/5.json")
 
-    json_second = chain_fitted_after.save_chain("data/3.json")
+    json_expected = chain_fitted_after.save_chain("data/7.json")
 
-    for i in range(1, 4):
-        os.remove(f"data/{i}.json")
-
-    delete_fitted_models(json.loads(json_first))
-    assert json_first == json_second
-
-
-def test_absolute_relative_paths_correctly_no_exception():
-    chain = create_chain()
-    chain.save_chain("data/test/1.json")
-
-    absolute_path = os.path.join(os.path.abspath("data/2.json"))
-    chain.save_chain(absolute_path)
-
-    chain.load_chain("data/test/1.json")
-    chain.load_chain(absolute_path)
-
-    os.remove("data/test/1.json")
-    os.remove(absolute_path)
-    os.rmdir("data/test")
+    assert json_actual == json_expected
 
 
 def test_import_custom_json_object_to_chain_and_fit_correctly_no_exception():
@@ -253,37 +256,14 @@ def test_import_custom_json_object_to_chain_and_fit_correctly_no_exception():
     train_data = InputData.from_csv(train_file_path)
 
     data_path = str(os.path.dirname(__file__))
-    json_file_path = os.path.join(data_path, '..', 'data', 'test_custom_json_template.json')
+    json_file_path = os.path.join(data_path, "data", "test_custom_json_template.json")
 
     chain = Chain()
     chain_template = ChainTemplate(chain)
     chain_template.import_from_json(json_file_path)
 
     chain.fit(train_data)
-    json_actual = chain.save_chain("data/1.json")
-
-    delete_fitted_models(json.loads(json_actual))
-    os.remove("data/1.json")
-
-
-def create_four_depth_chain():
-    knn_node = PrimaryNode('knn')
-    lda_node = PrimaryNode('lda')
-    xgb_node = PrimaryNode('xgboost')
-    logit_node = PrimaryNode('logit')
-
-    logit_node_second = SecondaryNode('logit', nodes_from=[knn_node, lda_node])
-    xgb_node_second = SecondaryNode('xgboost', nodes_from=[logit_node])
-
-    qda_node_third = SecondaryNode('qda', nodes_from=[xgb_node_second])
-    knn_node_third = SecondaryNode('knn', nodes_from=[logit_node_second, xgb_node])
-
-    knn_root = SecondaryNode('knn', nodes_from=[qda_node_third, knn_node_third])
-
-    chain = Chain()
-    chain.add_node(knn_root)
-
-    return chain
+    chain.save_chain("data/8.json")
 
 
 def test_extract_subtree_root():
@@ -306,17 +286,26 @@ def test_extract_subtree_root():
 
 
 def test_atomized_chain_import_export_correctly():
-    chain = create_chain_with_several_chain_models_nested()
-    chain.save_chain("data/chain_model_1.json")
-
     train_file_path, test_file_path = get_scoring_case_data_paths()
     train_data = InputData.from_csv(train_file_path)
 
-    chain = Chain()
-    chain.load_chain("data/chain_model_1.json")
+    chain = create_chain_with_several_nested_atomized_model()
     chain.fit(train_data)
-    json_actual = chain.save_chain("data/check_load_model_chain_1.json")
+    json_expected = chain.save_chain("data/atomized_model_1.json")
 
-    delete_fitted_models(json.loads(json_actual))
-    os.remove("data/chain_model_1.json")
-    os.remove("data/check_load_model_chain_1.json")
+    chain = Chain()
+    chain.load_chain("data/atomized_model_1.json")
+
+    with open("data/atomized_model_1.json", 'r') as json_file:
+        json_to_del = json.load(json_file)
+        delete_fitted_models(json_to_del)
+
+        dir_path = os.path.abspath("data/atomized_model_1")
+        if os.path.exists(dir_path):
+            shutil.rmtree(dir_path)
+
+        os.remove("data/atomized_model_1.json")
+
+    json_actual = chain.save_chain("data/atomized_model_1.json")
+
+    assert json_actual == json_expected
